% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{bindProteins}
\alias{bindProteins}
\title{bindProteins}
\usage{
bindProteins(proteinGraph, speedDate, bindingAffinity, propSpeedDating = 0.75,
  bindingLength = 5)
}
\arguments{
\item{proteinGraph}{is an igraph object specifying the protein-protein bindings}

\item{speedDate}{is the matrix of cells that should speed date}

\item{bindingAffinity}{is the matrix of binding affinities for the different proteins}

\item{propSpeedDating}{proportion of speed dating pairs that should actually bind}

\item{bindingLength}{number of timesteps in which the highest affinity proteins should bind}
}
\value{
\code{igraph} a proteinGraph object with new protein-protein bindings specified
}
\description{
the bindProteins function
}
\examples{


}
