% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{cellAggregationBarplot}
\alias{cellAggregationBarplot}
\title{cellAggregationBarplot}
\usage{
cellAggregationBarplot(cellAggregationResult)
}
\arguments{
\item{cellAggregationResult}{result object obtained from running cellAggregator()}
}
\value{
\code{ggplot} ggplot object of barplot of number of edges of each type per timestep
}
\description{
the cellAggregationBarplot function
}
\examples{

numCells = c(25,25)
numProtsPerCell = rbind(c(5,0,0),c(5,5,5))
bindingAffinity = cbind(c(1,0,0), c(0,1,0), c(0,0,1))

cellAggregationResult = cellAggregator(
 numCells,
 numProtsPerCell,
 bindingAffinity,
 timesteps = 100,
 burnIn = 0.75,
 verbose = TRUE,
 includeListOfGraphs = TRUE,
 plot = TRUE
)

cellAggregationResult$t_index
unlist(cellAggregationResult$listoftIndex)
cellAggregationBarplot(cellAggregationResult)

}
