% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{cellAggregator}
\alias{cellAggregator}
\title{cellAggregator}
\usage{
cellAggregator(numCells, numProtsPerCell, bindingAffinity, timesteps = 100,
  burnIn = 0.75, verbose = TRUE, includeListOfGraphs = TRUE,
  plot = FALSE, ...)
}
\arguments{
\item{numCells}{vector of the number of cells per cell population}

\item{numProtsPerCell}{matrix (number of cell populations x number of protein types) specifying number of proteins per cell}

\item{bindingAffinity}{matrix of binding affinities between proteins (should be symmetric matrix)}

\item{timesteps}{(default 100) number of timesteps}

\item{burnIn}{(default 0.75) consider the last burnIn proportion of timesteps for index calculation}

\item{verbose}{(default TRUE) print messages}

\item{includeListOfGraphs}{(default TRUE) include list of graphs in results object}

\item{plot}{(default FALSE) if TRUE prints a breakdown of the input parameters}

\item{...}{arguments go to bindProteins()}

\item{cellGraph}{is an igraph object specifying the cell-cell bindings (only one of proteinGraph or cellGraph should be specified)}
}
\value{
\code{list} cellAggregationResult object containing the input parameters: numCells, numProtsPerCell, bindingAffinity, timesteps, burnIn, verbose, includeListOfGraphs; and output objects: listofCellGraphs (list of cell graphs per timestep), listofProteinGraphs (list of protein graphs per timestep), listoftIndex (list of t index per timestep), listofTabulation (list of tabulation of proteins per timestep), and t_index (final t index summary value across entire simulation).
}
\description{
the cellAggregator function
}
\examples{

numCells = c(25,25)
numProtsPerCell = rbind(c(5,0,0),c(5,5,5))
bindingAffinity = cbind(c(1,0,0), c(0,1,0), c(0,0,1))

cellAggregationResult = cellAggregator(
 numCells,
 numProtsPerCell,
 bindingAffinity,
 timesteps = 100,
 burnIn = 0.75,
 verbose = TRUE,
 includeListOfGraphs = TRUE,
 plot = TRUE
)

cellAggregationResult$t_index
unlist(cellAggregationResult$listoftIndex)
cellAggregationBarplot(cellAggregationResult)

}
