% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{countDown}
\alias{countDown}
\title{bindProteins}
\usage{
countDown(proteinGraph)
}
\arguments{
\item{proteinGraph}{is an igraph object specifying the protein-protein bindings}
}
\value{
\code{igraph} a proteinGraph object with a countdown on the number of timesteps remaining
}
\description{
the bindProteins function
}
\examples{


}
