% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{getproteinStatus}
\alias{getproteinStatus}
\title{getproteinStatus}
\usage{
getproteinStatus(proteinGraph)
}
\arguments{
\item{proteinGraph}{is an igraph object specifying the protein-protein bindings}
}
\value{
\code{vector} character vector of "Bound" or "Unbound"
}
\description{
the getproteinStatus function
}
\examples{


}
