% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{mixingIndex}
\alias{mixingIndex}
\title{mixingIndex}
\usage{
mixingIndex(proteinGraph = NULL, cellGraph = NULL)
}
\arguments{
\item{proteinGraph}{is an igraph object specifying the protein-protein bindings (only one of proteinGraph or cellGraph should be specified)}

\item{cellGraph}{is an igraph object specifying the cell-cell bindings (only one of proteinGraph or cellGraph should be specified)}
}
\value{
\code{list} object containing the t-index and tabulation of number of edges between each pair of proteins
}
\description{
the mixingIndex function
}
\examples{


}
