% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{proteinGraphToCellGraph}
\alias{proteinGraphToCellGraph}
\title{proteinGraphToCellGraph}
\usage{
proteinGraphToCellGraph(proteinGraph, forPlotting = TRUE)
}
\arguments{
\item{proteinGraph}{is an igraph object specifying the protein-protein bindings}

\item{forPlotting}{optimise the igraph object for plotting (default TRUE, can be set to FALSE for speed)}
}
\value{
\code{igraph} cellGraph object network of cell-cell bindings
}
\description{
the proteinGraphToCellGraph function
}
\examples{


}
