% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{speedDateCells}
\alias{speedDateCells}
\title{speedDateCells}
\usage{
speedDateCells(cellGraph)
}
\arguments{
\item{cellGraph}{is an igraph object specifying the cell-cell bindings}
}
\value{
\code{matrix} a set of cell pairs which should speed date
}
\description{
the speedDateCells function
}
\examples{


}
