% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAggregator.R
\name{generateCellPopulation}
\alias{generateCellPopulation}
\title{generateCellPopulation}
\usage{
generateCellPopulation(numCells, numProtsPerCell, plot = TRUE)
}
\arguments{
\item{numCells}{is a vector of length N_cellpop containing the number of cells each should contain}

\item{numProtsPerCell}{is a N_cellpop x N_prot matrix containing the number of each type of protein appearing in each cellpop (rows are cell populations and columns are proteins).}

\item{plot}{logical if you should plot the cell information}
}
\value{
\code{igraph} output of this function is the igraph network object for protein network, without any edges
}
\description{
the generateCellPopulation function
}
\examples{

numCells = c(25,25,25)
numProtsPerCell = rbind(c(5,5,5),c(5,5,5), c(5,5,5))
proteinGraph = generateCellPopulation(numCells,numProtsPerCell)
plot(proteinGraph)

}
